Update: move server mocks to a separate package.
